refactor: 시리즈 PanelModule 컴포넌트 사용
