+++ src/components/sidebar/PanelModule.tsx
